测试发现某个等级之后int不够了

+++ src/main/java/com/gto/gtocore/api/playerskill/SkillValues.java
@@ -8,6 +8,6 @@ public class SkillValues {
 
     public static class ExperienceIncome {
 
-        public static final int EAT_MEAT = 10;
+        public static final long EAT_MEAT = 10;
     }
 }
